create home screen and add it in app js file
